Add the indeed-web-scraper.py program
Adds the indeed web scraper that scrapes through a web page, specifically
indeed. It turns its content including the title of the job, the company,
the location, and summary. Content was added to README.md to give context
to the program and its abilities and limitations. It also includes additions
that can be implemented later.